move all functions related to benchmark analysis to separate file
